mhlo: fix build errors and warnings in tosa pipeline
Prior to this patch, the MHLO tosa directory was skipped in the CMake
builds, and some of the include directives pointed to incorrect header
files.  This patch fixes the build rules and the header paths so that
the mhlo-tosa-opt binary can be built using CMake.

+++ tensorflow/compiler/xla/mlir_hlo/tosa/transforms/legalize_mhlo/legalize_mhlo.cc
@@ -17,6 +17,7 @@ limitations under the License.
 #include <utility>
 
 #include "mhlo/IR/hlo_ops.h"
+#include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Tosa/IR/TosaOps.h"
 #include "mlir/IR/BuiltinAttributes.h"
 #include "mlir/Parser/Parser.h"
@@ -24,7 +25,7 @@ limitations under the License.
 #include "transforms/passes.h"
 
 #define GEN_PASS_DEF_TOSALEGALIZEMHLOPASS
-#include "transforms/passes.h.inc"
+#include "tosa/transforms/passes.h.inc"
 
 #define PASS_NAME "tosa-legalize-mhlo"
 #define DEBUG_TYPE PASS_NAME
@@ -293,7 +294,7 @@ struct ConvertMhloGatherOp : public OpRewritePattern<mhlo::GatherOp> {
 
     auto startIndexMap = op.getDimensionNumbers().getStartIndexMap();
     for (const auto& startIndex : llvm::enumerate(startIndexMap)) {
-      if (startIndex.value() != startIndex.index()) {
+      if (startIndex.value() != static_cast<int64_t>(startIndex.index())) {
         return rewriter.notifyMatchFailure(op,
                                            "start_index_map must be in order");
       }
